refactor: implement callback pattern for key commands
- Extract KeyCommandHandler from editor to separate class
- Move command ID constants
- Implement callbacks for program navigation
